requirements.txt
django
djangorestframework
python-decouple
djangorestframework-simplejwt
pillow
twilio
phonenumbers
django-jazzmin
drf-material  
drf-spectacular  
drf-spectacular-sidecar
icecream
uvicorn
gunicorn
whitenoise
django-debug-toolbar
django-cors-headers
django-extensions
django-import-export[all]
eskiz_sms
environs